Update TagBot.yml
The tag bot was not updating the releases anymore as, I think, there were some changes in github actions or somewhere. This should fix it (copied from one of my projects), but make sure that the project has `secrets.DOCUMENTER_KEY` set up correctly. I cannot change this as I do not have rights to this repo...